Update pages-methods.php
To get only the user accessible pages for the menu (in template files) with easy chaining like so

```
$page->children()->accessibleByCurrentUser()->visible()
```

I created another function for the pages object, that works directly with the current user role. As there is no role for a not logged in user and Kirby returns false if you ask for the user, I had to temporarily create one. In my code, I even created an extension for Kirby $site object with a function $site->userRole() that does lines 48-55 and returns role wether there is a logged in user or not.